requirements.txt
Django
celery
whitenoise
cython
numpy
Pillow
djangorestframework
markdown
django-filter
